use unsigned long for IdType (to code more than 67 Mio elements)
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-spgrid/trunk@488 f45f52a6-3457-4210-b8e0-e9d642ddcde3

+++ dune/grid/spgrid/grid.hh
@@ -97,10 +97,8 @@ namespace Dune
 
       typedef SPGlobalIdSet< const Grid > GlobalIdSet;
       typedef SPLocalIdSet< const Grid > LocalIdSet;
-      //typedef typename GlobalIdSet::IdType GlobalIdType;
-      //typedef typename LocalIdSet::IdType LocalIdType;
-      typedef unsigned int GlobalIdType;
-      typedef unsigned int LocalIdType;
+      typedef unsigned long GlobalIdType;
+      typedef unsigned long LocalIdType;
 
       template< int codim >
       struct Codim
